python/requirements.txt
Flask
Flask-Cors
Flask-MySQLdb
PyJWT
bcrypt
